The texture cache manager mixed mask and regular bitmaps when tracking memory consumption, leading to wrong assumptions of available memory.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@969 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/bm.c
@@ -1067,7 +1067,7 @@ for (i = gameData.pig.tex.nBitmaps [0], bmP = gameData.pig.tex.bitmaps [0] + i;
 	OglFreeBmTexture (bmP);
 	if (bmP->bmTexBuf) {
 		D2_FREE (bmP->bmTexBuf);
-		UseBitmapCache (bmP, -bmP->bmProps.h * bmP->bmProps.rowSize);
+		UseBitmapCache (bmP, (int) -bmP->bmProps.h * (int) bmP->bmProps.rowSize);
 		}
 	}
 gameData.pig.tex.nExtraBitmaps = gameData.pig.tex.nBitmaps [0];
